Move initial guess functions to input.py
