+++ colabfold/batch.py
@@ -68,7 +68,9 @@ from colabfold.input import (
     pair_msa,
     msa_to_str,
     get_queries,
-    safe_filename
+    safe_filename,
+    modified_mapping,
+    pdb_to_string,
 )
 from colabfold.relax import relax_me
 from colabfold.alphafold import extra_ptm
@@ -166,25 +168,6 @@ def validate_and_fix_mmcif(cif_file: Path):
         shutil.copy2(cif_file, str(cif_file) + ".bak")
         with open(cif_file, "a") as f:
             f.write(CIF_REVISION_DATE)
-
-modified_mapping = {
-  "MSE" : "MET", "MLY" : "LYS", "FME" : "MET", "HYP" : "PRO",
-  "TPO" : "THR", "CSO" : "CYS", "SEP" : "SER", "M3L" : "LYS",
-  "HSK" : "HIS", "SAC" : "SER", "PCA" : "GLU", "DAL" : "ALA",
-  "CME" : "CYS", "CSD" : "CYS", "OCS" : "CYS", "DPR" : "PRO",
-  "B3K" : "LYS", "ALY" : "LYS", "YCM" : "CYS", "MLZ" : "LYS",
-  "4BF" : "TYR", "KCX" : "LYS", "B3E" : "GLU", "B3D" : "ASP",
-  "HZP" : "PRO", "CSX" : "CYS", "BAL" : "ALA", "HIC" : "HIS",
-  "DBZ" : "ALA", "DCY" : "CYS", "DVA" : "VAL", "NLE" : "LEU",
-  "SMC" : "CYS", "AGM" : "ARG", "B3A" : "ALA", "DAS" : "ASP",
-  "DLY" : "LYS", "DSN" : "SER", "DTH" : "THR", "GL3" : "GLY",
-  "HY3" : "PRO", "LLP" : "LYS", "MGN" : "GLN", "MHS" : "HIS",
-  "TRQ" : "TRP", "B3Y" : "TYR", "PHI" : "PHE", "PTR" : "TYR",
-  "TYS" : "TYR", "IAS" : "ASP", "GPL" : "LYS", "KYN" : "TRP",
-  "CSD" : "CYS", "SEC" : "CYS"
-}
-
-order_to_restype = {v: k for k, v in residue_constants.restype_order_with_x.items()}
 
 class ReplaceOrRemoveHetatmSelect(Select):
   def accept_residue(self, residue):
@@ -310,60 +293,6 @@ def pad_input(
         num_templates=4,
     )  # template_mask (4, 4) second value
     return input_fix
-
-
-def pdb_to_string(
-        pdb_file: str, 
-        chains: Optional[str] = None, 
-        models: Optional[list] = None
-    ) -> str:
-  '''read pdb file and return as string'''
-
-  if chains is not None:
-    if "," in chains: chains = chains.split(",")
-    if not isinstance(chains,list): chains = [chains]
-  if models is not None:
-    if not isinstance(models,list): models = [models]
-
-  modres = {**modified_mapping}
-  lines = []
-  seen = []
-  model = 1
-  
-  if "\n" in pdb_file:
-    old_lines = pdb_file.split("\n")
-  else:
-    with open(pdb_file,"rb") as f:
-      old_lines = [line.decode("utf-8","ignore").rstrip() for line in f]  
-  for line in old_lines:
-    if line[:5] == "MODEL":
-      model = int(line[5:])
-    if models is None or model in models:
-      if line[:6] == "MODRES":
-        k = line[12:15]
-        v = line[24:27]
-        if k not in modres and v in residue_constants.restype_3to1:
-          modres[k] = v
-      if line[:6] == "HETATM":
-        k = line[17:20]
-        if k in modres:
-          line = "ATOM  "+line[6:17]+modres[k]+line[20:]
-      if line[:4] == "ATOM":
-        chain = line[21:22]
-        if chains is None or chain in chains:
-          atom = line[12:12+4].strip()
-          resi = line[17:17+3]
-          resn = line[22:22+5].strip()
-          if resn[-1].isalpha(): # alternative atom
-            resn = resn[:-1]
-            line = line[:26]+" "+line[27:]
-          key = f"{model}_{chain}_{resn}_{resi}_{atom}"
-          if key not in seen: # skip alternative placements
-            lines.append(line)
-            seen.append(key)
-      if line[:5] == "MODEL" or line[:3] == "TER" or line[:6] == "ENDMDL":
-        lines.append(line)
-  return "\n".join(lines)
 
 class file_manager:
     def __init__(self, prefix: str, result_dir: Path):
@@ -626,27 +555,6 @@ def predict_structure(
     return {"rank":rank,
             "metric":metric,
             "result_files":result_files}
-
-def decode_structure_sequences(
-    aatype_array: List[int], 
-    chain_index_array: List[int], 
-    order_dict: Dict[int, str] = order_to_restype
-) -> List[str]:
-    decoded_sequences = []
-    current_sequence = []
-
-    for i in range(len(aatype_array)):
-        amino_acid = order_dict[aatype_array[i]]
-        if i == 0 or chain_index_array[i] == chain_index_array[i - 1]:
-            current_sequence.append(amino_acid)
-        else:
-            decoded_sequences.append("".join(current_sequence))
-            current_sequence = [amino_acid]
-
-    # Append the last sequence
-    decoded_sequences.append("".join(current_sequence))
-
-    return decoded_sequences
 
 def get_msa_and_templates(
     jobname: str,
